requirements.txt
streamlit>=1.29   #Frontend Framework
supabase>=2.0.2 #DB Client
fastapi>=0.104.1   #Handles endpoints
uvicorn[standard]>=0.24.0 # ASGI server for FastAPI
python-dotenv>=1.0.0 
pydantic